Arreglando ultimos errores

+++ src/main/java/ec/edu/espe/distribuidas/matricula/service/MatriculaService.java
@@ -45,13 +45,13 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Slf4j
 public class MatriculaService {
-
+    
     private final MatriculaRepository matriculaRepository;
     private final EstudianteRepository estudianteRepository;
     private final CursoRepository cursoRepository;
     private final PeriodoRepository periodoRepository;
     private final DetalleMatriculaRepository detalleMatriculaRepository;
-
+    
     public MatriculaService(MatriculaRepository matriculaRepository, EstudianteRepository estudianteRepository, CursoRepository cursoRepository, PeriodoRepository periodoRepository, DetalleMatriculaRepository detalleMatriculaRepository) {
         this.matriculaRepository = matriculaRepository;
         this.estudianteRepository = estudianteRepository;
@@ -59,34 +59,7 @@ public class MatriculaService {
         this.periodoRepository = periodoRepository;
         this.detalleMatriculaRepository = detalleMatriculaRepository;
     }
-
-    private DetalleMatricula verificarHorario(List<String> errorCursos, List<DetalleMatricula> detalleMatriculas, Matricula matricula, Curso curso, Integer creditos) {
-        for (DetalleMatricula detalleMatricula : detalleMatriculas) {
-            for (Horario horarioActual : curso.getHorarios()) {
-                List<Horario> horariosChoque = detalleMatricula.getCurso().getHorarios()
-                        .stream().filter(horario -> horario.getDia().equals(horarioActual.getDia())
-                        && horario.getHoraInicio().after(horarioActual.getHoraFin()) && horario.getHoraFin().before(horarioActual.getHoraInicio()))
-                        .collect(Collectors.toList());
-                if (!horariosChoque.isEmpty()) {
-                    log.error("El curso:" + curso.getNrc() + " se choca con uno o mas cursos actuales");
-                    errorCursos.add("El curso:" + curso.getNrc() + " se choca con uno o mas cursos actuales");
-                    return null;
-                }
-            }
-        }
-
-        DetalleMatricula detalleMatriculaNueva = new DetalleMatricula();
-        detalleMatriculaNueva.setMatricula(matricula);
-        detalleMatriculaNueva.setCurso(curso);
-        detalleMatriculaNueva.setEstado("REP");
-        detalleMatriculaNueva.setFecha(new Date());
-        creditos += curso.getAsignatura().getCreditos();
-        log.info("Creditos:{}", creditos);
-        curso.setCupo((short) (curso.getCupo() - 1));
-        //log.info("Nueva materia:{}",detalleMatriculaNueva.getCurso().getAsignatura().getNombre());
-        return detalleMatriculaNueva;
-    }
-
+    
     @Transactional
     public List<String> matricularse(MatriculaRQ matriculaRQ) {
         Matricula matricula;
@@ -101,37 +74,41 @@ public class MatriculaService {
             }
         }
         List<DetalleMatricula> detalleMatriculas = new ArrayList<>();
-
+        
         Optional<Estudiante> estudianteOpt = this.estudianteRepository.findByCorreo(matriculaRQ.getCorreo());
         if (estudianteOpt.isEmpty()) {
             throw new EntityNotFoundException("No se encontro el estudiante con el correo: " + matriculaRQ.getCorreo());
         }
         Estudiante estudiante = estudianteOpt.get();
-
+        
         Optional<Periodo> periodo = this.periodoRepository.findById(matriculaRQ.getPeriodo());
         if (periodo.isEmpty()) {
             throw new EntityNotFoundException("No se encontro el periodo con el ID" + matriculaRQ.getPeriodo());
         }
         Integer creditos = 0;
+        if (matricula.getCodigo() != null) {
+            creditos = matricula.getCreditosTotales().intValue();
+        }
+        
         List<String> errorCursos = new ArrayList<>();
-
+        
         for (Integer crs : matriculaRQ.getCursos()) {
             Optional<Curso> cursoOpt = this.cursoRepository.findById(crs);
-
+            
             if (cursoOpt.isEmpty()) {
                 errorCursos.add("El curso con el ID: " + crs + " no existe");
                 continue;
             }
             Curso curso = cursoOpt.get();
-
+            
             if (matricula.getCodigo() != null) {
                 List<DetalleMatricula> cursosRepetidos = matricula.getDetalle().stream().filter(d -> d.getCurso().getAsignatura().getNombre().equals(curso.getAsignatura().getNombre())).collect(Collectors.toList());
                 if (!cursosRepetidos.isEmpty()) {
-                    errorCursos.add("ya esta matriculado en el curso con el NRC: " + curso.getNrc());
+                    errorCursos.add("ya esta matriculado en una materia de " + curso.getAsignatura().getNombre());
                     continue;
                 }
             }
-
+            
             log.info("curso periodo:{}", curso.getPeriodo());
             log.info("periodo matricula:_{}", periodo.get());
             if (!curso.getPeriodo().equals(periodo.get())) {
@@ -148,15 +125,15 @@ public class MatriculaService {
                 log.error("El curso:" + curso.getNrc() + " no cuenta con cupos disponibles");
                 continue;
             }
-
+            
             boolean carrera = false;
             boolean prerequisito = false;
             int nPre = 0;
-
+            
             List<DetalleMatricula> materiasEstudiante = this.detalleMatriculaRepository.findByEstudiante(estudiante.getCodigo());
-
+            
             log.info("Detalles OP:{}", materiasEstudiante);
-
+            
             for (Prerequisito pre : curso.getAsignatura().getPrerequisitos()) {
                 for (DetalleMatricula detMatricula : materiasEstudiante) {
                     if (pre.getPrerequisito().equals(detMatricula.getCurso().getAsignatura())) {
@@ -164,7 +141,7 @@ public class MatriculaService {
                         break;
                     }
                 }
-
+                
             }
             if (curso.getAsignatura().getPrerequisitos().size() == nPre) {
                 prerequisito = true;
@@ -172,7 +149,7 @@ public class MatriculaService {
                 log.error("No cumplio con los prerequisitos");
                 errorCursos.add("No cumple con los prerequisitos para el curso:" + curso.getNrc());
             }
-
+            
             for (CarreraCurso carreraCurso : curso.getCarreraCursos()) {
                 log.info("Carrera verificacion:{}", carreraCurso.getCarrera().getNombre());
                 if (carreraCurso.getCarrera().equals(estudiante.getCarrera())) {
@@ -184,7 +161,7 @@ public class MatriculaService {
                 log.error("Carrera no disponible");
                 errorCursos.add("Su carrera no esta contemplada en el curso:" + curso.getNrc());
             }
-
+            
             boolean ver = false;
             if ((curso.getAsignatura().getPrerequisitos().isEmpty() || prerequisito) && (curso.getCarreraCursos().isEmpty() || carrera)) {
                 for (DetalleMatricula detalleMatricula : detalleMatriculas) {
@@ -212,16 +189,16 @@ public class MatriculaService {
                     detalleMatriculaNueva.setFecha(new Date());
                     creditos += curso.getAsignatura().getCreditos();
                     log.info("Creditos:{}", creditos);
-                    curso.setCupo((short) (curso.getCupo() - 1));
-
+                    curso.setDisponible((short) (curso.getDisponible() - 1));
+                    
                     detalleMatriculas.add(detalleMatriculaNueva);
                 }
-
+                
             }
         }
-
+        
         matricula.setEstudiante(estudiante);
-
+        
         log.info("CreditosDes:{}", creditos);
         log.info("Errores:{}", errorCursos);
         matricula.setCreditosTotales(BigDecimal.valueOf(creditos));
@@ -239,41 +216,47 @@ public class MatriculaService {
             errorCursos.add("No se logro matricular en ninguna materia");
             throw new MatriculaConflictException("No se logro matricular en ninguna materia", errorCursos);
         }
-
+        
     }
-
+    
     public Matricula buscarMatricula(String correo, Integer periodo) {
-
+        
         Optional<Estudiante> estudianteOpt = this.estudianteRepository.findByCorreo(correo);
         if (estudianteOpt.isEmpty()) {
             throw new EntityNotFoundException("No se encontro el estudiante con el correo" + correo);
         }
         Estudiante estudiante = estudianteOpt.get();
-
+        
         Optional<Periodo> per = this.periodoRepository.findById(periodo);
         if (per.isEmpty()) {
             throw new EntityNotFoundException("No se encontro el periodo con el ID" + periodo);
         }
-
+        
         Optional<Matricula> matriculaOpt = this.matriculaRepository.findByEstudianteAndPeriodo(estudiante, per.get());
         if (matriculaOpt.isEmpty()) {
             throw new EntityNotFoundException("No se encontro la una matricula en el periodo: " + per.get().getNombre() + " del estudiante " + estudiante.getApellido() + " " + estudiante.getNombre());
         }
-
+        
         return matriculaOpt.get();
     }
-
+    
     public void borrarDetalleMatricula(Integer id) {
+        Optional<DetalleMatricula> detalleMatriculaOpt = this.detalleMatriculaRepository.findById(id);
+        if (detalleMatriculaOpt.isEmpty()) {
+            throw new EntityNotFoundException("No se encontro el detalle");
+        }
+        detalleMatriculaOpt.get().getMatricula().setCreditosTotales(detalleMatriculaOpt.get().getMatricula().getCreditosTotales().subtract(BigDecimal.valueOf(detalleMatriculaOpt.get().getCurso().getAsignatura().getCreditos())));
+        detalleMatriculaOpt.get().getCurso().setDisponible((short) (detalleMatriculaOpt.get().getCurso().getDisponible() + 1));
         this.detalleMatriculaRepository.deleteById(id);
     }
-
+    
     public List<Matricula> obtenerMatriculasPorEstudiante(String correo) {
         Optional<Estudiante> estudianteOpt = this.estudianteRepository.findByCorreo(correo);
         if (estudianteOpt.isEmpty()) {
             throw new EntityNotFoundException("No se encontro el estudiante con el correo: " + correo);
         }
         Estudiante estudiante = estudianteOpt.get();
-
+        
         return estudiante.getMatricula();
     }
 }
